refactoring api dirs structure
